WIP on getting OperationViewer into storybook
